Build workarounds for various kernels
CentOS uses kernel 4.18 and in that version, CC_HAVE_ASM_GOTO is defined
based on the script from scripts/gcc-goto.sh using gcc, which is not
relevant for our build with clang, so let's use CC_HAVE_ASM_GOTO=1 as a
workaround.

The CentOS kernel from
https://vault.centos.org/8-stream/BaseOS/Source/SPackages/kernel-4.18.0-348.el8.src.rpm
has some differences with the upstream one:

> ifdef CONFIG_RETPOLINE
> ifeq ($(RETPOLINE_CFLAGS),)
> $(error You are building kernel with non-retpoline compiler, please update your compiler.)
> endif

So let's use 'RETPOLINE_CFLAGS=-DRETPOLINE_CFLAGS' as a workaround.